Infinite RadioGatún: Computation cost, where we have a curious property: O(n) (the time needed to calculate an infinitely long string, where n is the mill size) will always take longer than an O(n * n) (the time needed to calculate a short string, where n is the mill size) calculation.
Of course, it takes (l * n) + (n * n) calculations to come up with a
given hash, where l is the length of the input string and n is the
mill size, so for the blank rounds to not matter in determining how
long it takes to calculate the hash, l has to always be an order of
magnitude larger than n.

So, it's not linear: "l" has to get bigger too for us to keep the blank
rounds from using a singificant amount of calculation time!