tinyrg32.c and pwgen.sh improved; better passwords!
tinyrg32 can now generate secure alphanumeric passwords with a user-chosen
level of entropy.  This allows the user to make their own security vs.
difficulty to memorize trade off for passwords.  60 bits should deter
anyone except the most determined attacker; 80 bits should be able to
deter pretty much anyone.

Also, pwgen.sh is now smart enough to compile tinyrg32 if it is not in
the user’s path (it uses a name like tinyrg32-123.c to minimize the chance
of deleting a needed file)