nanorg32.c & nanorg32-vertical.c I have not only made the code for both of these smaller (in the case of nanorg32-vertical.c, have less actual code in the file), but I have fixed the problem with shifts on non-x86 CPUs.
Let’s look at this example C program:
#include <stdio.h>
#include <stdint.h>
int main() {
        uint32_t a;
        for(a=0;a<64;a++) {
                printf("%d %d\n",a,(-a%32));
        }
}
Here, we get, for the second column, 0,31,30,29,28,27,26,25,24
If we had "int a" instead of using the unsigned type, we would have
0,-1,-2,-3, etc.
By using the unsigned type, the modulous operator is guaranteed to
give us a positive number.  Note that this code will technically
give us undefined behavior: We take whatever "-a" becomes when
we make it a positive unsigned integer (which is CPU implemention
dependent), then run the modulo operation on it.  It’s only a
“Euclidian modulo” operator when the amount we modulo by is a power
of 2 (such as 32) on machines which use twos compliment for negative
integers (read: All modern CPUs).