Lua/rg32.c: We do not need rand32 or randomstrseed
We do not need randomstrseed() because of Lua’s built in coercion.
So, I have removed rg32.randomseed, renamed randomstrseed randomseed,
and have it accept a string argument.  If it’s given a number, Lua
will make the number a string for us.

We can use two calls to rand16 to get rand32’s results, and do not have
to deal with the pesky issue of maybe someone’s using 32-bit floats
with only 24 binary digits of mantissa.