rg32-infected: Let’s allow only the third word in the bell to change
Let’s create in infinite Radio Gatún:

- Belt size: Smaller unit in sexy prime pair
- Mill size: Larger unit in sexy prime pair
- Initial state: Top row of the mill can have any initial value
- Word size: 32 bits
- Belt rows: 3

In more detail, we allow any value for all but the first two and final
two words in the top row of the belt.  So, for standard RadioGatún[32],
which has a belt row size of 13, we allow the middle nine words to have
any 32-bit value, giving us 2^288 possible different RadioGatún[32]
initial states.

We have a minimum of 16 blank rounds, and we have [Mill width - 3] blank
rounds. In addition, If the mill width is greater than 40, we divide
the mill width by 40, rounding down. The resulting number is added to
the number of blank rounds we have.

Given all this, let’s, presuming we have an infinite number of sexy
primes, make the belt and mill infinitely long. How many possible
InfiniteRadioGatún[32] states do we have?

The same as the quantity of real numbers:

* Each InfiniteRadioGatun[32] is infinite in size
* So let’s make a list of every single possible InfiniteRadioGatun[32] state
* We can now use the diagonal proof to make a new state on the list.  Add 1
  (mod 2^32) to the first number of the first row, the second number of the
  second row, the third number of the third row and so on to create a new
  number. This number will not be in the list, hence this is not a
  “countable” infinity or InfiniteRadioGatún[32] states.

The code change ensures that a one-bit change to the third word of the top
mill completely changes the derived hash value.